+++ src/components/main/navigation/NavigationBar.js
@@ -5,18 +5,20 @@ import { FiMenu, FiX, FiSearch } from "react-icons/fi";
 import { MdShoppingBasket } from "react-icons/md";
 import logoImg from "img/logo.png";
 import { rootMenu } from "api/rootMenu";
+import SubNavigationBar from "./SubNavigationBar";
 
 function NavigationBar(props) {
   const [opend, setOpen] = useState(false); //false = bars, true = times
-  const handleClick = () => {
+  const [clickedMenu, setClickedMenu] = useState("");
+  const handleOnFixMenu = () => {
     setOpen(!opend);
   };
-  const resetClick = () => {
+  const clickOnMenu = (e) => {
     setOpen(false);
   };
   const menuList = rootMenu.map((menu, index) => (
     <li key={index} className={classes.navItem}>
-      <Link to={menu.key} className={classes.navLink} onClick={resetClick}>
+      <Link to={menu.key} className={classes.navLink} onClick={clickOnMenu}>
         {menu.label}
       </Link>
     </li>
@@ -27,13 +29,12 @@ function NavigationBar(props) {
   return (
     <>
       <nav className={classes.navbar}>
-        <Link to="/" onClick={resetClick}>
+        <Link to="/" onClick={clickOnMenu}>
           <img className={classes.navLogo} src={logoImg} alt="로고" />
         </Link>
         {!opend && (
           <ul className={classes.navLinks}>
             <div className={classes.navLeft}>{menuList}</div>
-
             <div className={classes.navRight}>
               <li className={classes.navItem}>
                 <form>
@@ -51,7 +52,7 @@ function NavigationBar(props) {
                 <Link
                   to="/signup"
                   className={classes.navLink}
-                  onClick={resetClick}
+                  onClick={clickOnMenu}
                 >
                   회원가입
                 </Link>
@@ -60,7 +61,7 @@ function NavigationBar(props) {
                 <Link
                   to="/signin"
                   className={classes.navLink}
-                  onClick={resetClick}
+                  onClick={clickOnMenu}
                 >
                   로그인
                 </Link>
@@ -73,22 +74,31 @@ function NavigationBar(props) {
             </div>
           </ul>
         )}
-        <div className={classes.navIcon} onClick={handleClick}>
+        <div className={classes.navIcon} onClick={handleOnFixMenu}>
           {opend ? <FiX /> : <FiMenu />}
         </div>
       </nav>
+      <SubNavigationBar />
 
       {opend && (
         <ul className={classes.navLinks.active}>
           {menuList}
           {/* 추후, 로그인 정보 불러와서 조건 걸 예정 */}
           <li className={classes.navItem}>
-            <Link to="/signup" className={classes.navLink} onClick={resetClick}>
+            <Link
+              to="/signup"
+              className={classes.navLink}
+              onClick={clickOnMenu}
+            >
               회원가입
             </Link>
           </li>
           <li className={classes.navItem}>
-            <Link to="/signin" className={classes.navLink} onClick={resetClick}>
+            <Link
+              to="/signin"
+              className={classes.navLink}
+              onClick={clickOnMenu}
+            >
               로그인
             </Link>
           </li>